Renaming only: MapConfigurator -> MapProvider; BaseLayerSource -> MapConfigurator.
