+++ collect_app/src/main/java/org/odk/collect/android/activities/GeoPointMapActivity.java
@@ -24,7 +24,7 @@ import android.widget.ImageButton;
 import android.widget.TextView;
 
 import org.odk.collect.android.R;
-import org.odk.collect.android.map.MapConfigurator;
+import org.odk.collect.android.map.MapProvider;
 import org.odk.collect.android.map.MapFragment;
 import org.odk.collect.android.map.MapPoint;
 import org.odk.collect.android.preferences.MapsPreferences;
@@ -125,7 +125,7 @@ public class GeoPointMapActivity extends BaseGeoMapActivity {
         zoomButton = findViewById(R.id.zoom);
 
         Context context = getApplicationContext();
-        MapConfigurator.createMapFragment(context)
+        MapProvider.createMapFragment(context)
             .addTo(this, R.id.map_container, this::initMap, this::finish);
     }
